Fix all PR comments
