+++ src/types/auth.ts
@@ -1,0 +1,7 @@
+export interface ILoginUser {
+  email: string;
+  password: string;
+}
+export interface ISignupUser extends ILoginUser {
+  name: string;
+}
