+++ src/types/models.ts
@@ -4,14 +4,6 @@ export interface IUser {
   _id: string;
   months?: IMonth[];
 }
-export interface ILoginUser {
-  email: string;
-  password: string;
-}
-export interface ISignupUser extends ILoginUser {
-  name: string;
-}
-
 export interface IIncome {
   _id: string;
   category: String;
