+++ src/utils/app.constants.ts
@@ -1,0 +1,8 @@
+export const APP = {
+  name: 'Budget manager',
+  currency: '€',
+  localStorage: {
+    authToken: 'authToken',
+  },
+  projectApi: `http://localhost:5005/api`,
+};
